requirements.txt
streamlit
pdfplumber
python-docx
markdown
langchain
langchain-community
sentence-transformers
faiss-cpu
pyyaml
numpy